// src/app/interface/interface.ts
export interface User {
    name: string;
    age: number;
    gender: string;
    dob: string;
    qualification: string;
    place: string;
    designation: string;
    experience: string;
    mail: string;
    phoneno: number;
    image: string;
    status: string;
    coverletter: string;
    date: number;
    id: string;
  }
  